// convex/lib/zepto/generated/core.ts
export interface paths {
// ...
    "/payments": {
        parameters: {
            query?: never;
            header?: {
                /** @description API version in YYYYMMDD format. Defaults to 20250101 (legacy) when omitted. */
                "Zepto-API-Version"?: components["parameters"]["ZeptoAPIVersion"];
            };
            path?: never;
            cookie?: never;
        };
        /** List all Payments */
        get: operations["ListAllPayments"];
        put?: never;
        /**
         * Make a Payment
         * @description To enable custom payment flows, the required payment channel can be selected by setting the _channel_ attribute to one of the following combinations:
         *
         *     <ul>
         *       <li>["new_payments_platform"] - for faster payments 24/7/365</li>
         *       <li>["direct_entry"] - for slower traditional payments</li>
         *       <li>["new_payments_platform", "direct_entry"] - enables automatic channel switching if a payment fails on the NPP</li>
         *     </ul>
         *     <aside class="notice">We now require supplying an <code>Idempotency-Key</code> header when performing this request to ensure you can safely retry the action in case of an issue. If the header value is different to one provided previously, we will be treating a request as a new operation which may lead to duplicate payments. To understand more on how to make idempotent requests, please refer to our <a href="https://docs.zeptopayments.com/v20260101/reference/idempotent-requests">Idempotent requests guide</a>.</aside>
         */
        post: operations["MakeAPayment"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/payment_requests": {
        parameters: {
            query?: never;
            header?: {
                /** @description API version in YYYYMMDD format. Defaults to 20250101 (legacy) when omitted. */
                "Zepto-API-Version"?: components["parameters"]["ZeptoAPIVersion"];
            };
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /**
         * Request Payment
         * @description <aside class="notice">We now require supplying an <code>Idempotency-Key</code> header when performing this request to ensure you can safely retry the action in case of an issue. If the header value is different to one provided previously, we will be treating a request as a new operation which may lead to a duplicate funds collection. To understand more on how to make idempotent requests, please refer to our <a href="https://docs.zeptopayments.com/v20260101/reference/idempotent-requests">Idempotent requests guide</a>.</aside>
         */
        post: operations["MakeAPaymentRequest"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/transfers": {
        parameters: {
            query?: never;
            header?: {
                /** @description API version in YYYYMMDD format. Defaults to 20250101 (legacy) when omitted. */
                "Zepto-API-Version"?: components["parameters"]["ZeptoAPIVersion"];
            };
            path?: never;
            cookie?: never;
        };
        /** List all Transfers (Available soon) */
        get: operations["ListAllTransfers"];
        put?: never;
        /**
         * Add a Transfer
         * @description Use this endpoint when you want to create a Transfer between any 2 of your float/bank accounts.
         *     <aside class="notice">We now require supplying an <code>Idempotency-Key</code> header when performing this request to ensure you can safely retry the action in case of an issue. If the header value is different to one provided previously, we will be treating a request as a new operation which may lead to duplicate payments. To understand more on how to make idempotent requests, please refer to our <a href="https://docs.zeptopayments.com/v20260101/reference/idempotent-requests">Idempotent requests guide</a>.</aside>
         */
        post: operations["AddATransfer"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}